PR commits

+++ src/app/import/components/modal-upload/modal-upload.component.ts
@@ -1,8 +1,7 @@
-import { AfterViewInit, Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { Component, ElementRef, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
 import { createDefaultModalUploadModal, getSubtitle, getTitle, UploadModals } from './modal-upload-helper';
 import { UploadFileType, UploadOptions } from '../helpers/upload-types';
 import { UploadComponent } from '../upload/upload.component';
-import { UploadHelper } from '../helpers/upload-helper';
 import { UploadStates } from '../../services/s3.enums';
 import { interval } from 'rxjs';
 import { NotificationService } from 'src/app/base/services/notification.service';
@@ -12,12 +11,12 @@ import { NotificationService } from 'src/app/base/services/notification.service'
   templateUrl: './modal-upload.component.html',
   styleUrls: ['./modal-upload.component.scss']
 })
-export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
+export class ModalUploadComponent implements OnInit, OnChanges, OnDestroy {
 
   @Input() uploadFileType: UploadFileType;
   @Input() isModalOpen: boolean;
   @Input() uploadOptions: UploadOptions;
-  @Input() projectId?: string;
+  @Input() projectId: string;
 
   @Output() closeModalEvent = new EventEmitter();
   @Output() fileAttached = new EventEmitter<File>();
@@ -35,13 +34,15 @@ export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
   @ViewChild(UploadComponent) baseComponent: UploadComponent;
   uploadModals: UploadModals = createDefaultModalUploadModal();
   file: File | null = null;
-  uploadHelper: UploadHelper;
   title: string;
   subTitle: string;
 
   constructor() { }
 
   ngOnInit(): void {
+    this.title = getTitle(this.uploadFileType);
+    this.subTitle = getSubtitle(this.uploadFileType);
+
     NotificationService.subscribeToNotification(this, {
       projectId: this.projectId,
       whitelist: ['file_upload'],
@@ -49,21 +50,21 @@ export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
     });
   }
 
-  ngAfterViewInit(): void {
-    this.uploadHelper = new UploadHelper(this.baseComponent, this.baseComponent.recordNewUploadHelper, this.baseComponent.recordAddUploadHelper, this.baseComponent.existingProjectUploadHelper, this.baseComponent.lookupListsUploadHelper);
+  ngOnDestroy(): void {
+    NotificationService.unsubscribeFromNotification(this, this.projectId);
   }
 
   ngOnChanges(changes: SimpleChanges): void {
     if (changes.isModalOpen) {
       this.uploadModals.uploadFile.open = changes.isModalOpen.currentValue;
-      this.title = getTitle(this.uploadFileType);
-      this.subTitle = getSubtitle(this.uploadFileType);
     }
   }
 
   closeModal(): void {
     this.uploadModals.uploadFile.open = false;
-    this.closeModalEvent.emit();
+    if (this.uploadOptions.closeModalOnClick) {
+      this.closeModalEvent.emit();
+    }
     if (this.uploadOptions.knowledgeBaseId) {
       this.refetchLookupLists.emit(true);
     }
@@ -74,7 +75,7 @@ export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
     this.baseComponent.uploadOptions = this.uploadOptions;
     if (this.projectId) this.baseComponent.projectId = this.projectId;
     this.uploadModals.uploadFile.doingSomething = true;
-    this.uploadHelper.upload(this.file);
+    this.baseComponent.uploadHelper.upload(this.file);
   }
 
   setFile(file: File): void {
@@ -83,7 +84,7 @@ export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
 
   optionClicked(button: string): void {
     if (button == 'CLOSE') this.closeModal();
-    if (button == 'ACCEPT') {
+    else if (button == 'ACCEPT') {
       const x = interval(1000).subscribe(() => {
         if (this.baseComponent.progressState?.progress === 100 && this.baseComponent.progressState?.state === UploadStates.DONE) {
           this.baseComponent.resetUpload();
@@ -97,7 +98,7 @@ export class ModalUploadComponent implements OnInit, AfterViewInit, OnChanges {
 
   handleWebsocketNotification(msgParts: string[]): void {
     if (msgParts[1] === 'file_upload' && msgParts[4] === UploadStates.ERROR) {
-      this.baseComponent.deleteExistingProject();
+      if (this.uploadOptions.deleteProjectOnFail) this.baseComponent.deleteExistingProject();
     }
   }
 }
